requirements.txt
asgiref==3.3.4
dj-config-url==0.1.1
Django==3.2.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
gunicorn==20.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
